requirements.txt
pandas>=2.0.0
tkinter
PyQt6>=6.4.0
pyinstaller>=5.7.0
Pillow>=10.0.0 